Changed shapes .h to .hpp
