+++ zaitsev.vladimir/T1/Parallelogram.cpp
@@ -1,4 +1,4 @@
-#include "Parallelogram.h"
+#include "parallelogram.hpp"
 #include <stdexcept>
 #include <cstring>
 #include <cmath>
